requirements.txt
absl-py==0.10.0
aniso8601==8.0.0
appnope==0.1.0
astunparse==1.6.3
attrs==19.3.0
awscli==1.16.313
backcall==0.1.0
bar-chart-race==0.1.0
bleach==3.1.4
bokeh==2.1.1
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.1
colorcet==2.0.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
firebase-admin==4.4.0
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
flask-restplus==0.13.0
gast==0.3.3
gcloud==0.17.0
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
holoviews==1.13.3
httplib2==0.18.1
hvplot==0.6.0
idna==2.10
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jws==0.1.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauth2client==3.0.0
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
panel==0.9.7
param==1.9.3
parso==0.6.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyct==0.4.6
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.6.1
pymongo==3.11.0
pyparsing==2.4.6
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
Pyrebase==3.0.27
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2019.3
pyviz-comms==0.7.6
PyYAML==5.2
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==3.4.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
statsmodels==0.11.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0
